refactor(ui): collocate schemas with corresponding elements
