+++ apps/demo/src/app/layout.tsx
@@ -2,7 +2,7 @@ import { contentPath, initContent } from '@/content'
 import { DeepDishProvider } from '@deepdish/cms'
 import { createJsonResolver } from '@deepdish/resolvers/json'
 import { configure } from '@deepdish/ui/config'
-import { typographySchema } from '@deepdish/ui/schemas'
+import { typographySchema } from '@deepdish/ui/typography'
 import type { Metadata } from 'next'
 import localFont from 'next/font/local'
 import './globals.css'
