+++ packages/ui/src/elements/link.tsx
@@ -1,8 +1,16 @@
 import 'server-only'
 
+import { z } from 'zod'
 import { DeepDish } from '../deepdish'
-import type { LinkValue } from '../schemas'
 import type { ElementProps } from '../types'
+
+export const linkSchema = z.object({
+  destination: z.string().optional(),
+  href: z.string().optional(),
+  title: z.string().optional(),
+})
+
+type LinkValue = z.infer<typeof linkSchema>
 
 export function Link(props: ElementProps<'a', string>) {
   return (
