+++ packages/ui/src/elements/media.tsx
@@ -1,8 +1,15 @@
 import 'server-only'
 
+import { z } from 'zod'
 import { DeepDish } from '../deepdish'
-import type { AudioValue, ImageValue, VideoValue } from '../schemas'
 import type { ElementProps } from '../types'
+
+export const audioSchema = z.object({
+  fallback: z.string().optional(),
+  source: z.string().optional(),
+})
+
+type AudioValue = z.infer<typeof audioSchema>
 
 export function Audio(props: ElementProps<'audio', string>) {
   return (
@@ -22,6 +29,14 @@ export function Audio(props: ElementProps<'audio', string>) {
     />
   )
 }
+
+export const imageSchema = z.object({
+  description: z.string().optional(),
+  source: z.string().optional(),
+  title: z.string().optional(),
+})
+
+type ImageValue = z.infer<typeof imageSchema>
 
 export function Image(props: ElementProps<'img'>) {
   return (
@@ -45,6 +60,13 @@ export function Image(props: ElementProps<'img'>) {
     />
   )
 }
+
+export const videoSchema = z.object({
+  fallback: z.string().optional(),
+  source: z.string().optional(),
+})
+
+type VideoValue = z.infer<typeof videoSchema>
 
 export function Video(props: ElementProps<'video', string>) {
   return (
