+++ packages/ui/src/elements/typography.tsx
@@ -1,9 +1,13 @@
 import 'server-only'
 
+import { z } from 'zod'
 import { type ContentFormat, sanitizeContent } from '../content'
 import { DeepDish } from '../deepdish'
-import type { TypographyValue } from '../schemas'
 import type { ElementProps, IntrinsicElement } from '../types'
+
+export const typographySchema = z.string()
+
+type TypographyValue = z.infer<typeof typographySchema>
 
 type TypographyProps<E extends IntrinsicElement> = ElementProps<E, string> & {
   format?: ContentFormat
